refactor: move permissions to project level directory
